Actually, never mind that. I'm going to freeze it.
I don't have GHC 7.6, and I don't actually think it's terribly important
that we support it. Anyone who has 7.6 who wants to take ownership of
making things build for 7.6 can do so, but it's utter insanity to not
freeze one's dependencies (even during development) if "reproducible
builds" is at all a desired property.

By that token, I'm only targeting 7.8 on travis; if someone wants to
figure out how to freeze a different set of dependencies for 7.6 and
hook that up with travis, they are welcome to.